refactor: Replace order-service with orders
