--- order-service/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/orderservice/OrderServiceApplication.java
@@ -1,13 +1,0 @@
-package com.github.itsAkshayDubey.eventdrivenarchitecture.orderservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class OrderServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(OrderServiceApplication.class, args);
-	}
-
-}
