--- order-service/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/orderservice/core/entity/Order.java
@@ -1,33 +1,0 @@
-package com.github.itsAkshayDubey.eventdrivenarchitecture.orderservice.core.entity;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-import com.github.itsAkshayDubey.eventdrivenarchitecture.orderservice.core.model.OrderStatus;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table(name = "orders")
-public class Order {
-	
-	@Id
-	@Column(unique = true)
-    private String orderId;
-    private String productId;
-    private String userId;
-    private int quantity;
-    private String addressId;
-    @Enumerated(EnumType.STRING)
-    private OrderStatus orderStatus;
-
-}
