--- order-service/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/orderservice/core/errorhandling/ErrorMessage.java
@@ -1,17 +1,0 @@
-package com.github.itsAkshayDubey.eventdrivenarchitecture.orderservice.core.errorhandling;
-
-import java.util.Date;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class ErrorMessage {
-	
-	private String message;
-	private Date timestamp;
-
-}
