--- order-service/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/orderservice/core/events/OrderApprovedEvent.java
@@ -1,12 +1,0 @@
-package com.github.itsAkshayDubey.eventdrivenarchitecture.orderservice.core.events;
-
-import com.github.itsAkshayDubey.eventdrivenarchitecture.orderservice.core.model.OrderStatus;
-
-import lombok.Value;
-
-@Value
-public class OrderApprovedEvent {
-
-	private final String orderId;
-	private final  OrderStatus orderStatus = OrderStatus.APPROVED;
-}
