--- order-service/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/orderservice/core/events/OrderCreatedEvent.java
@@ -1,15 +1,0 @@
-package com.github.itsAkshayDubey.eventdrivenarchitecture.orderservice.core.events;
-
-import com.github.itsAkshayDubey.eventdrivenarchitecture.orderservice.core.model.OrderStatus;
-
-import lombok.Data;
-
-@Data
-public class OrderCreatedEvent {
-	private String orderId;
-	private String productId;
-	private String userId;
-	private int quantity;
-	private String addressId;
-	private OrderStatus orderStatus;
-}
