--- order-service/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/orderservice/core/events/OrderRejectEvent.java
@@ -1,13 +1,0 @@
-package com.github.itsAkshayDubey.eventdrivenarchitecture.orderservice.core.events;
-
-import com.github.itsAkshayDubey.eventdrivenarchitecture.orderservice.core.model.OrderStatus;
-
-import lombok.Value;
-
-@Value
-public class OrderRejectEvent {
-
-	private final String orderId;
-	private final String message;
-	private OrderStatus orderStatus = OrderStatus.REJECTED;
-}
