--- order-service/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/orderservice/core/model/OrderStatus.java
@@ -1,7 +1,0 @@
-package com.github.itsAkshayDubey.eventdrivenarchitecture.orderservice.core.model;
-
-public enum OrderStatus {
-	
-	CREATED, APPROVED, REJECTED
-
-}
