--- order-service/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/orderservice/core/model/OrderSummary.java
@@ -1,12 +1,0 @@
-package com.github.itsAkshayDubey.eventdrivenarchitecture.orderservice.core.model;
-
-import lombok.Value;
-
-@Value
-public class OrderSummary {
-
-	
-	private final String orderId;
-	private final OrderStatus orderStatus;
-	private final String message;
-}
