--- order-service/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/orderservice/core/repo/OrderRepo.java
@@ -1,11 +1,0 @@
-package com.github.itsAkshayDubey.eventdrivenarchitecture.orderservice.core.repo;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.github.itsAkshayDubey.eventdrivenarchitecture.orderservice.core.entity.Order;
-
-public interface OrderRepo extends JpaRepository<Order, String>{
-	
-	Order findByOrderId(String orderId);
-
-}
