--- order-service/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/orderservice/query/FindOrderQuery.java
@@ -1,10 +1,0 @@
-package com.github.itsAkshayDubey.eventdrivenarchitecture.orderservice.query;
-
-import lombok.Value;
-
-@Value
-public class FindOrderQuery {
-
-	private final String orderId;
-	
-}
