--- order-service/src/test/java/com/github/itsAkshayDubey/eventdrivenarchitecture/orderservice/OrderServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.github.itsAkshayDubey.eventdrivenarchitecture.orderservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class OrderServiceApplicationTests {
-
-//	@Test
-//	void contextLoads() {
-//	}
-
-}
