+++ orders/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/orders/Orders.java
@@ -1,0 +1,13 @@
+package com.github.itsAkshayDubey.eventdrivenarchitecture.orders;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Orders {
+
+	public static void main(String[] args) {
+		SpringApplication.run(Orders.class, args);
+	}
+
+}
