+++ orders/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/orders/command/ApproveOrderCommand.java
@@ -1,0 +1,14 @@
+package com.github.itsAkshayDubey.eventdrivenarchitecture.orders.command;
+
+import org.axonframework.modelling.command.TargetAggregateIdentifier;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class ApproveOrderCommand {
+	
+	@TargetAggregateIdentifier
+	private String orderId;
+}
