+++ orders/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/orders/command/RejectOrderCommand.java
@@ -1,0 +1,16 @@
+package com.github.itsAkshayDubey.eventdrivenarchitecture.orders.command;
+
+import org.axonframework.modelling.command.TargetAggregateIdentifier;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class RejectOrderCommand {
+	
+	@TargetAggregateIdentifier
+	private final String orderId;
+	private final String message;
+	
+}
