+++ orders/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/orders/core/entity/Order.java
@@ -1,0 +1,33 @@
+package com.github.itsAkshayDubey.eventdrivenarchitecture.orders.core.entity;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+import com.github.itsAkshayDubey.eventdrivenarchitecture.orders.core.model.OrderStatus;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "orders")
+public class Order {
+	
+	@Id
+	@Column(unique = true)
+    private String orderId;
+    private String productId;
+    private String userId;
+    private int quantity;
+    private String addressId;
+    @Enumerated(EnumType.STRING)
+    private OrderStatus orderStatus;
+
+}
