+++ orders/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/orders/core/events/OrderCreatedEvent.java
@@ -1,0 +1,15 @@
+package com.github.itsAkshayDubey.eventdrivenarchitecture.orders.core.events;
+
+import com.github.itsAkshayDubey.eventdrivenarchitecture.orders.core.model.OrderStatus;
+
+import lombok.Data;
+
+@Data
+public class OrderCreatedEvent {
+	private String orderId;
+	private String productId;
+	private String userId;
+	private int quantity;
+	private String addressId;
+	private OrderStatus orderStatus;
+}
