+++ orders/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/orders/core/model/OrderStatus.java
@@ -1,0 +1,7 @@
+package com.github.itsAkshayDubey.eventdrivenarchitecture.orders.core.model;
+
+public enum OrderStatus {
+	
+	CREATED, APPROVED, REJECTED
+
+}
