+++ orders/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/orders/core/model/OrderSummary.java
@@ -1,0 +1,12 @@
+package com.github.itsAkshayDubey.eventdrivenarchitecture.orders.core.model;
+
+import lombok.Value;
+
+@Value
+public class OrderSummary {
+
+	
+	private final String orderId;
+	private final OrderStatus orderStatus;
+	private final String message;
+}
