+++ orders/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/orders/query/FindOrderQuery.java
@@ -1,0 +1,10 @@
+package com.github.itsAkshayDubey.eventdrivenarchitecture.orders.query;
+
+import lombok.Value;
+
+@Value
+public class FindOrderQuery {
+
+	private final String orderId;
+	
+}
