+++ orders/src/test/java/com/github/itsAkshayDubey/eventdrivenarchitecture/orders/OrdersTests.java
@@ -1,0 +1,13 @@
+package com.github.itsAkshayDubey.eventdrivenarchitecture.orders;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class OrdersTests {
+
+//	@Test
+//	void contextLoads() {
+//	}
+
+}
